requirements.txt
numpy
Pillow
ultralytics
torch
torchvision
torchaudio
pandas
tqdm
scipy
requests
PyYAML
opencv-python
confluent-kafka
flask
Werkzeug
keras_preprocessing
pyanime4k
wget
patool
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
tensorflow
scikit-learn